Move contracts under ubuntupro/contracts
